is it now

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -14,7 +14,8 @@ if __name__ == "__main__":
     Session = sessionmaker(bind=engine)
     session = Session()
 
-    for instance in session.query(State).filter(State.name.like('%a%')):
-        session.delete(instance)
+    for instance in session.query(State).filter(State.name):
+        if "a" in instance.name:
+            session.delete(instance)
 
     session.commit()
